Rename 'generic_table' to 'base_values_table' in base schema
Renamed the 'generic_table' to 'base_values_table' to clarify its use in the base schema. This name change is reflected in associated comments, SQL queries, and actions in the base_schema.kf file. This aims to improve code readability and understanding of the table's purpose – to serve as a basic structure for every database.