Replace float weights with integer weights in truflation
This commit replaces the previous method of using float numbers as weights for streams in the truflation module with integer weights. It also moves the Fraction function to the utils package and adjusts its usages. Now, weights need to be input as integers and the sum no longer needs to be 1. This avoids precision and deterministic issues related to float numbers and simplifies the usage by removing the need for sum validation. Example: Instead of inputting '0.1' for 10%, you now enter '10'.