feat: compute index out of two other indexes (#52)
* work in progress creating schema and improving extension

* Add support for range values in BaseStreamExt

The code now supports getting range values in BaseStreamExt by adding a getOrDefault function which sets default values if none are provided. It also modifies methods index and value by implementing a range value option, providing more dynamic and customizable functionality for users. This includes error checking for expected number of values.

* Delete outdated meat_poultry_fish_eggs schema alternatives

Four old schema alternatives for 'meat_poultry_fish_eggs' have been removed. These schemas were not in use and needed to be cleaned up from the codebase for better maintainability and clarity. This included two alternative schemas for table and index weights, a single table schema, and a multiple-table schema.

* Add transform_source.py for CSV file processing

* Remove from_base.go and corresponding test file

* Update base schema and remove admin check

* Add Kwil data query in experiment commands

* Remove 'from_base' extension

* Improve date validation in basestream.go

* Update basestream to handle range and multiple values

In this commit, changes have been made to support the retrieval of a range of values in the basestream module. The mockScalar method now supports multiple results and the index and value methods can now handle date ranges.

* Reorder and add debugging instruction in experiment.md

* Update result return process in base_schema and basestream with workaround

* Update stream paths and error messages

Updated the stream paths in the 'experiment_table.kf' file to start with a forward slash. Improved the function to handle the DBID from a path or a DBID string in the 'stream.go' file. Also, fixed an error message to correctly indicate the number of expected inputs.

* Update experiment.md with new database commands

Two empty lines were removed for brevity. Added new sections for deploying the 'experiment_table' and querying it. This involves calling the database with specific dates and parameters for the 'experiment_table'.

* Add command to drop experiment table before deployment

* Replace simple stream with weighted stream

* simplify csv samples

* Refactor weighted stream calculation

* Update query date in kwil-cli commands to reflect recent csv changes

* Add ascending order to sqlGetRangeValue query

The sqlGetRangeValue function has been updated to sort the results in ascending order. This change has been reflected in both the SQL statement itself and the related sprintf function call.

* Add utility function to derive DBID from path and utilize it in weighted stream

A utility function that gets the DBID from a path or a DBID has been added. This function has been then incorporated within the weighted stream module to replace an existing similar function, thereby reducing redundancy. An associated unit test has also been added to verify the functionality of this new utility function. In the weighted stream module, rounding to the nearest integer has also been added for the final value computation.

* Remove unused math imports from experiment_table.kf

* Update experiment scripts and reorganize resources

Revised the experiment guide and scripts in the develop_experiments directory. This update makes it easier to build, deploy and query experimental trials with Kwil. Relocated and renamed some resources for better organization, and made subtle data changes to allow easy verification of correct weights in projections.

* Rename 'generic_table' to 'base_values_table' in base schema

Renamed the 'generic_table' to 'base_values_table' to clarify its use in the base schema. This name change is reflected in associated comments, SQL queries, and actions in the base_schema.kf file. This aims to improve code readability and understanding of the table's purpose – to serve as a basic structure for every database.

* Update database deploy path in develop_scripts.md

The path for deploying the base schema in the develop_scripts.md file has been corrected. The previous relative path led to the wrong location, so it was updated to point to the correct location of the base_schema.kf file. This affects both the truflation_us_hotel_price and yahoo_finance_corn_futures databases.

* Add .gitignore to test_samples directory

* Refactor weighted_stream to compose_streams

In the truflation extension, the weighted_stream component was renamed to compose_streams. This refactoring reflects the more complex nature of the component that now supports composing multiple streams. Usage in experiments, 'composed.kf', was updated accordingly, and extensions are properly re-registered under the new name.

* Add tests for CalculateWeightedResultsWithFn and refactor code

New tests have been introduced to evaluate the CalculateWeightedResultsWithFn function under various scenarios. Refactoring within the compose_streams.go file has also been conducted to enhance readability and to isolate the CalculateWeightedResultsWithFn function for easier testing.

* Update error handling in basestream.go

* Update comment and position in basestream.go

* Update documentation with expected error scenarios

* Add rounding utility function and apply to weightMap values

A function named RoundToDecimalPlaces has been added into the utils.go file to handle rounding of float values to the desired number of decimal places. This new function is now utilized in the compose_streams.go file to standardize the values added to weightMap to three decimal places.

* Update RoundToDecimalPlaces function in utils.go

* update documentation comments on composed.kf

* Update documentation code for extensions

The commit comprehensively refactors code for the 'basestream', 'composed_1', and 'composed_2' extensions, impacting how they are used and called. Major changes include modifications to function inputs and structure. In parallel, the README documentation has been revised to reflect these code changes, providing updates and additional context where necessary.

* Refactor README.md removing accidental wraps

* Refactor README.md removing accidental wraps

* Replace float weights with integer weights in truflation

This commit replaces the previous method of using float numbers as weights for streams in the truflation module with integer weights. It also moves the Fraction function to the utils package and adjusts its usages. Now, weights need to be input as integers and the sum no longer needs to be 1. This avoids precision and deterministic issues related to float numbers and simplifies the usage by removing the need for sum validation. Example: Instead of inputting '0.1' for 10%, you now enter '10'.

* Update Kwil Node run command in dev script