Add count of schemas left to deploy in logs
The commit integrates a feature to display the remaining numbers of primitive and composed schemas left to be deployed in the deployment logs. This update allows the database deploy script to provide more informative and helpful feedback during the schemas deployment process. Additionally, the .gitignore file was updated to exclude a new file "all_tables.csv".