feat(raw sources): Add script that help pulling from DB directly (#56)
* Add data extraction and processing scripts

The commit includes two new scripts: `pull_db_data.py` to extract sample data from a database and store it in csv format, and `transform_raw_from_db.sh` that cleans the data, multiplying all values by 1000 and storing them into a temporary csv file. A .gitignore file was also added under the scripts directory to ensure sensitive database credentials and temporary csv files are not committed.

* commit first raw_from_db file as first demo

* fix .gitignore

* Add result format and dummy credentials in scripts

Added detailed notes in 'pull_db_data.py' about the format of the resulting CSV files generated by the script. Also added a new file 'secret_db_credentials.example.json' as a placeholder for database access credentials. This file serves as a guide and should be populated with real access information.

* Refactor comments into docstring in pull_db_data.py

The comments at the start of the pull_db_data.py script have been converted into a multiline docstring. This ensures that the information is included in Python's built-in help system, making accessibility and understanding of the code easier for developers. This includes details on the script's purpose, requirements, limitations, and the format of the result.