Refactor stream value handling in compose_streams
Introduced a flag 'should_discard' to improve the management of new values to be pushed in the stream. This effectively removes the need for a labelled outer continue statement. We're now also able to continue parsing that date to find if any stream have new values to assign as latest.