Add argparse functionality to source map script
This update introduces argparse functionality to the process_all.py script, allowing users to run the script with a `--minimum` argument for testing minimally. This argument lets the script to read a smaller and more manageable .csv file (`categories-tables-us-min.csv`) instead of the full one. Improvements have also been made to the deployment instructions.