Update cache key paths in CI workflow
The cache key paths in the continuous integration (CI) workflow have been updated. Previously, the paths were dependent on the OS of the runner, but now they are constant and set to `/tmp/buildx-cache-ci` and `/tmp/buildx-cache`. This change simplifies the access to the buildx cache across different operating systems.