Update gateway binary file path and remove smoke test
The binary file path in the `move` command has been updated to include an extra 'kgw' in its destination. This change aligns with the expected location in the docker-compose file. Moreover, the smoke test command that uses curl has been removed. This deletion eliminates an unnecessary operation in the AWS Nitro Enclaves CLI setup.