Implement use of ACM for SSL on CloudFront
The code now implements AWS Certificate Manager for SSL certificate generation. This change primarily impacts CloudFront, as ACM will now handle certificate generation for that service. Previously, certificate functionality was handled through CertBot, which was removed due to limitations with rate and security implications. An ACM certificate function was created for this purpose.