Add curl and jq to gateway Dockerfile and dynamically fetch chain_id
This update adds curl and jq commands to the Dockerfile in the gateway deployment, facilitating a dynamic fetch of the 'chain_id' from the tsn-db API. As a result of this change, the static 'chain_id' is removed from the 'kgw.base-config.json' file. This optimizes container set up by reducing hardcoded configurations and enhancing flexibility.