Add tsn-config.dockerfile and update deployment Dockerfiles
This commit adds a new file, tsn-config.dockerfile, to generate a configuration file for a single node or use an externally provided configuration. The Dockerfiles for deployments have also been updated, particularly revising the `ARG` variables and the application of the configuration path. Additionally, the filename `gen-kwil-conf.dockerfile` is changed to `network-conf.dockerfile`.