Update metadata handling in stream contract
This commit includes several updates to the handling of metadata in the composed stream contract. The insert_metadata procedure now supports multiple types of values. An error system is implemented to notify when invalid types are inputted. It also includes modification in get_metadata function and changing the current_block value from 1 to 2.