Refactor 'get_current_version' method to handle disabled taxonomies
The 'get_current_version' method has been updated to include an extra parameter which will allow it to handle disabled taxonomies. Depending on the boolean value of this parameter, the method will either return the version of a taxonomy that is not disabled or just the latest version of any taxonomy.